LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/block-view-timeline-current-time-vertical-rl.tentative-expected.txt


FAIL View timeline with container having vertical-rl layout assert_approx_equals: values do not match for "Timeline's currentTime at container start boundary" expected -150 +/- 0.125 but got 64.91228485107422
